+++ sickbeard/providers/__init__.py
@@ -27,7 +27,7 @@ from sickbeard import logger, encodingKludge as ek
 from . import newznab, omgwtfnzbs, womble
 # torrent
 from . import alpharatio, beyondhd, bithdtv, bitmetv, btn, filelist, freshontv, funfile, gftracker, grabtheinfo, \
-    hd4free, hdbits, hdspace, iptorrents, kat, morethan, pisexy, pretome, rarbg, scc, scenetime, shazbat, speedcd, \
+    hd4free, hdbits, hdspace, iptorrents, kat, morethan, pisexy, pretome, privatehd, rarbg, scc, scenetime, shazbat, speedcd, \
     thepiratebay, torrentbytes, torrentday, torrenting, torrentleech, torrentshack, transmithe_net, tvchaosuk
 # anime
 from . import anizb, nyaatorrents, tokyotoshokan
@@ -59,6 +59,7 @@ __all__ = ['omgwtfnzbs',
            'morethan',
            'pisexy',
            'pretome',
+           'privatehd',
            'rarbg',
            'scc',
            'scenetime',